osiris/requirements.txt
click
rich
colorama
pyfiglet
jupyter
pyTigerGraph
google-cloud-bigquery[bqstorage,pandas]
graphistry